Add program document sync. ab#16683
The codebase has undergone significant changes to incorporate the handling of program documents. A new `ProgramDocumentController` class and `ProgramDocumentService` have been added to manage operations related to program documents. The `Organization` class has been updated to include a `ProgramDocumentUrl` property and a `ProgramDocuments` collection. The `Badge` class has been updated to replace the `EducationProgramId` property with the `OrganizationId` property. A new `ProgramDocument` property has been added to the `AppDbContext` class, representing a database set of program documents. A new migration `AddProgramDocument` has been added to add the `ProgramDocument` table to the database.

The `EducationProgramMapperConfiguration` class has been renamed to `OrganizationMapperConfiguration` and a new `ProgramDocumentMapperConfiguration` class has been added for mapping between `ProgramDocument` and `ProgramDocumentDto` objects. The `IOrganizationService` interface has been updated to make the `AddOrganizationAsync` and `GetOrganizationAsync` methods nullable. The `Markdig` package, a Markdown processor for .NET, has been added to the `CrystallineSociety.Server.Api.csproj` project file.

The `IProgramDocumentService` interface was added to manage program documents. The `GitHubBadgeService` and `OrganizationService` classes were updated to include methods for retrieving program documents. The `ProgramDocumentService` class was added to implement the `IProgramDocumentService` interface. The `BadgeDto`, `OrganizationDto`, and `ProgramDocumentDto` classes were updated to include new properties. The `IServiceCollectionExtensions` class was updated to include the `IProgramDocumentUtilService` and `ProgramDocumentUtilService` services. The `IGitHubBadgeService` interface was updated to include a method for getting program documents. The `IProgramDocumentUtilService` interface was added to include methods for parsing and serializing program documents. The `BadgeUtilService` class was updated to include the `SpecJson` property in the `ParseBadge` method. The `ProgramDocumentUtilService` class was added to implement the `IProgramDocumentUtilService` interface.