Remove unnecessary steps in README.md
The most important changes made to the files are the addition of instructions on how to open and run a project using two different Integrated Development Environments (IDEs): Visual Studio Code and Visual Studio. These changes are aimed at providing clear instructions for users on how to get started with the project using either of the two recommended IDEs.

List of changes:

1. Added instructions on how to open and run a project using Visual Studio Code. This includes steps on downloading the software, opening the project folder, opening a new terminal within the IDE, and running the application using the `dotnet run` command.

2. Added instructions on how to open and run a project using Visual Studio. This includes steps on downloading the software and opening the project by selecting the `.sln` file of the project.

References to the code changes are not provided in the descriptions.